temp

+++ network/cvae_graph/dataloader.py
@@ -97,7 +97,7 @@ class GraphDataset(Dataset):
                         pooled_node_features[i, 2] /= 2
                         pooled_node_features[i, 3] /= 2
 
-            pooled_building_masks = torch.zeros((pooled_boundary_n + building_n))
+            pooled_building_masks = torch.zeros((pooled_boundary_n + building_n), dtype=torch.long)
             pooled_building_masks[pooled_boundary_n:] = 1
 
             if self.condition_type == 'image' or self.condition_type == 'image_resnet34':
